requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==4.4.2
Django==4.2
# Django==3.2.21
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
ffmpeg==1.4
frozenlist==1.3.3
gunicorn==21.2.0
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.9
inflection==0.5.1
jsonschema==4.17.3
jsonschema-specifications==0.1.3
# kombu==5.3.0b3
kombu==5.3.1
moviepy==1.0.3
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
packaging==23.1
Pillow==9.5.0
proglog==0.1.10
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.8.9
requests==2.31.0
# rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.5.0
yarl==1.9.2
django-allauth==0.58.1
